+++ contracts/solc-0.6/StringConstructor.sol
@@ -1,7 +1,12 @@
 pragma solidity ^0.6.0;
 
 contract StringConstructor {
-    constructor(string memory message) public {
+
+    modifier hasModifier() {
+        _;
+    }
+
+    constructor(string memory message) public hasModifier {
 
     }
 }