+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.3.6 (2022-01-31)
 
 - Remove calls to empty self unchained method from init method body, and empty parent unchained methods calls without parameters.
 - Persist modifiers on constructors.